AI Devs 3/MyApps/factory_analyzer-SOURCE/requirements.txt
openai>=1.0.0
python-dotenv>=0.19.0
requests>=2.26.0
Pillow>=8.3.1
python-magic>=0.4.24
tiktoken>=0.3.3 